refacor UT for pose transforms
