+++ mmaction/datasets/transforms/loading.py
@@ -4,7 +4,6 @@ import io
 import os
 import os.path as osp
 import shutil
-from typing import Dict
 
 import mmcv
 import numpy as np
@@ -263,174 +262,6 @@ class SampleFrames(BaseTransform):
                     f'twice_sample={self.twice_sample}, '
                     f'out_of_bound_opt={self.out_of_bound_opt}, '
                     f'test_mode={self.test_mode})')
-        return repr_str
-
-
-@TRANSFORMS.register_module()
-class UniformSampleFrames(BaseTransform):
-    """Uniformly sample frames from the video.
-
-    To sample an n-frame clip from the video. UniformSampleFrames basically
-    divide the video into n segments of equal length and randomly sample one
-    frame from each segment. To make the testing results reproducible, a
-    random seed is set during testing, to make the sampling results
-    deterministic.
-
-    Required Keys:
-
-        - total_frames
-        - start_index (optional)
-
-    Added Keys:
-
-        - frame_inds
-        - frame_interval
-        - num_clips
-        - clip_len
-
-    Args:
-        clip_len (int): Frames of each sampled output clip.
-        num_clips (int): Number of clips to be sampled. Defaults to 1.
-        test_mode (bool): Store True when building test or validation dataset.
-            Defaults to False.
-        seed (int): The random seed used during test time. Defaults to 255.
-    """
-
-    def __init__(self,
-                 clip_len: int,
-                 num_clips: int = 1,
-                 test_mode: bool = False,
-                 seed: int = 255) -> None:
-        self.clip_len = clip_len
-        self.num_clips = num_clips
-        self.test_mode = test_mode
-        self.seed = seed
-
-    def _get_train_clips(self, num_frames: int, clip_len: int) -> np.ndarray:
-        """Uniformly sample indices for training clips.
-
-        Args:
-            num_frames (int): The number of frames.
-            clip_len (int): The length of the clip.
-
-        Returns:
-            np.ndarray: The sampled indices for training clips.
-        """
-        all_inds = []
-        for clip_idx in range(self.num_clips):
-            if num_frames < clip_len:
-                start = np.random.randint(0, num_frames)
-                inds = np.arange(start, start + clip_len)
-            elif clip_len <= num_frames < 2 * clip_len:
-                basic = np.arange(clip_len)
-                inds = np.random.choice(
-                    clip_len + 1, num_frames - clip_len, replace=False)
-                offset = np.zeros(clip_len + 1, dtype=np.int32)
-                offset[inds] = 1
-                offset = np.cumsum(offset)
-                inds = basic + offset[:-1]
-            else:
-                bids = np.array(
-                    [i * num_frames // clip_len for i in range(clip_len + 1)])
-                bsize = np.diff(bids)
-                bst = bids[:clip_len]
-                offset = np.random.randint(bsize)
-                inds = bst + offset
-
-            all_inds.append(inds)
-
-        return np.concatenate(all_inds)
-
-    def _get_test_clips(self, num_frames: int, clip_len: int) -> np.ndarray:
-        """Uniformly sample indices for testing clips.
-
-        Args:
-            num_frames (int): The number of frames.
-            clip_len (int): The length of the clip.
-
-        Returns:
-            np.ndarray: The sampled indices for testing clips.
-        """
-
-        np.random.seed(self.seed)
-        all_inds = []
-        for i in range(self.num_clips):
-            if num_frames < clip_len:
-                start_ind = i if num_frames < self.num_clips \
-                    else i * num_frames // self.num_clips
-                inds = np.arange(start_ind, start_ind + clip_len)
-            elif clip_len <= num_frames < clip_len * 2:
-                basic = np.arange(clip_len)
-                inds = np.random.choice(
-                    clip_len + 1, num_frames - clip_len, replace=False)
-                offset = np.zeros(clip_len + 1, dtype=np.int64)
-                offset[inds] = 1
-                offset = np.cumsum(offset)
-                inds = basic + offset[:-1]
-            else:
-                bids = np.array(
-                    [i * num_frames // clip_len for i in range(clip_len + 1)])
-                bsize = np.diff(bids)
-                bst = bids[:clip_len]
-                offset = np.random.randint(bsize)
-                inds = bst + offset
-
-            all_inds.append(inds)
-
-        return np.concatenate(all_inds)
-
-    def transform(self, results: Dict) -> Dict:
-        """The transform function of :class:`UniformSampleFrames`.
-
-        Args:
-            results (dict): The result dict.
-
-        Returns:
-            dict: The result dict.
-        """
-        num_frames = results['total_frames']
-
-        if self.test_mode:
-            inds = self._get_test_clips(num_frames, self.clip_len)
-        else:
-            inds = self._get_train_clips(num_frames, self.clip_len)
-
-        inds = np.mod(inds, num_frames)
-        start_index = results.get('start_index', 0)
-        inds = inds + start_index
-
-        if 'keypoint' in results:
-            kp = results['keypoint']
-            assert num_frames == kp.shape[1]
-            num_person = kp.shape[0]
-            num_persons = [num_person] * num_frames
-            for i in range(num_frames):
-                j = num_person - 1
-                while j >= 0 and np.all(np.abs(kp[j, i]) < 1e-5):
-                    j -= 1
-                num_persons[i] = j + 1
-            transitional = [False] * num_frames
-            for i in range(1, num_frames - 1):
-                if num_persons[i] != num_persons[i - 1]:
-                    transitional[i] = transitional[i - 1] = True
-                if num_persons[i] != num_persons[i + 1]:
-                    transitional[i] = transitional[i + 1] = True
-            inds_int = inds.astype(np.int)
-            coeff = np.array([transitional[i] for i in inds_int])
-            inds = (coeff * inds_int + (1 - coeff) * inds).astype(np.float32)
-
-        results['frame_inds'] = inds.astype(np.int32)
-        results['clip_len'] = self.clip_len
-        results['frame_interval'] = None
-        results['num_clips'] = self.num_clips
-        return results
-
-    def __repr__(self):
-        repr_str = (f'{self.__class__.__name__}('
-                    f'clip_len={self.clip_len}, '
-                    f'num_clips={self.num_clips}, '
-                    f'test_mode={self.test_mode}, '
-                    f'seed={self.seed})')
         return repr_str
