+++ tests/datasets/transforms/test_sampling.py
@@ -402,90 +402,6 @@ class TestSampling(BaseTestLoading):
         assert np.max(sample_frames_results['frame_inds']) <= 40
         assert np.min(sample_frames_results['frame_inds']) >= 1
 
-    def test_uniform_sample_frames(self):
-        results = dict(total_frames=64, start_index=0)
-        sampling = UniformSampleFrames(
-            clip_len=8, num_clips=1, test_mode=True, seed=0)
-
-        assert repr(sampling) == ('UniformSampleFrames(clip_len=8, '
-                                  'num_clips=1, test_mode=True, seed=0)')
-        sampling_results = sampling(results)
-        assert sampling_results['clip_len'] == 8
-        assert sampling_results['frame_interval'] is None
-        assert sampling_results['num_clips'] == 1
-        assert_array_equal(sampling_results['frame_inds'],
-                           np.array([4, 15, 21, 24, 35, 43, 51, 63]))
-
-        results = dict(total_frames=15, start_index=0)
-        sampling = UniformSampleFrames(
-            clip_len=8, num_clips=1, test_mode=True, seed=0)
-        sampling_results = sampling(results)
-        assert sampling_results['clip_len'] == 8
-        assert sampling_results['frame_interval'] is None
-        assert sampling_results['num_clips'] == 1
-        assert_array_equal(sampling_results['frame_inds'],
-                           np.array([0, 2, 4, 6, 8, 9, 11, 13]))
-
-        results = dict(total_frames=7, start_index=0)
-        sampling = UniformSampleFrames(
-            clip_len=8, num_clips=1, test_mode=True, seed=0)
-        sampling_results = sampling(results)
-        assert sampling_results['clip_len'] == 8
-        assert sampling_results['frame_interval'] is None
-        assert sampling_results['num_clips'] == 1
-        assert_array_equal(sampling_results['frame_inds'],
-                           np.array([0, 1, 2, 3, 4, 5, 6, 0]))
-
-        results = dict(total_frames=7, start_index=0)
-        sampling = UniformSampleFrames(
-            clip_len=8, num_clips=8, test_mode=True, seed=0)
-        sampling_results = sampling(results)
-        assert sampling_results['clip_len'] == 8
-        assert sampling_results['frame_interval'] is None
-        assert sampling_results['num_clips'] == 8
-        assert len(sampling_results['frame_inds']) == 64
-
-        results = dict(total_frames=64, start_index=0)
-        sampling = UniformSampleFrames(
-            clip_len=8, num_clips=4, test_mode=True, seed=0)
-        sampling_results = sampling(results)
-        assert sampling_results['clip_len'] == 8
-        assert sampling_results['frame_interval'] is None
-        assert sampling_results['num_clips'] == 4
-        assert_array_equal(
-            sampling_results['frame_inds'],
-            np.array([
-                4, 15, 21, 24, 35, 43, 51, 63, 1, 11, 21, 26, 36, 47, 54, 56,
-                0, 12, 18, 25, 38, 47, 55, 62, 0, 9, 21, 25, 37, 40, 49, 60
-            ]))
-
-        results = dict(total_frames=64, start_index=0)
-        sampling = UniformSampleFrames(
-            clip_len=8, num_clips=1, test_mode=False, seed=0)
-        sampling_results = sampling(results)
-        assert sampling_results['clip_len'] == 8
-        assert sampling_results['frame_interval'] is None
-        assert sampling_results['num_clips'] == 1
-        assert len(sampling_results['frame_inds']) == 8
-
-        results = dict(total_frames=7, start_index=0)
-        sampling = UniformSampleFrames(
-            clip_len=8, num_clips=1, test_mode=False, seed=0)
-        sampling_results = sampling(results)
-        assert sampling_results['clip_len'] == 8
-        assert sampling_results['frame_interval'] is None
-        assert sampling_results['num_clips'] == 1
-        assert len(sampling_results['frame_inds']) == 8
-
-        results = dict(total_frames=15, start_index=0)
-        sampling = UniformSampleFrames(
-            clip_len=8, num_clips=1, test_mode=False, seed=0)
-        sampling_results = sampling(results)
-        assert sampling_results['clip_len'] == 8
-        assert sampling_results['frame_interval'] is None
-        assert sampling_results['num_clips'] == 1
-        assert len(sampling_results['frame_inds']) == 8
-
     def test_dense_sample_frames(self):
         target_keys = [
             'frame_inds', 'clip_len', 'frame_interval', 'num_clips',
